+++ .travis.yml
@@ -11,7 +11,6 @@ install:
   - ./.travis/install.sh
 script:
   - make lint
-  - make test-configs
   - make test
 after_success:
   - coveralls
