+++ config/engines/process_infinite.yaml
@@ -1,5 +1,5 @@
 engine: !engine
-  type: pnp.engines.process.ProcessEngine
+  type: pnp.engines.ProcessEngine
   queue_worker: 1
   retry_handler: !retry
     type: pnp.engines.SimpleRetryHandler
