+++ config/readme/01_hello_world.yaml
@@ -1,4 +1,3 @@
----
 - name: hello-world
   pull:
     plugin: pnp.plugins.pull.simple.Repeat
