+++ config/readme/06_unwrapping.yaml
@@ -1,4 +1,3 @@
----
 - name: unwrapping
   pull:
     plugin: pnp.plugins.pull.simple.Repeat
