+++ config/readme/07_unwrapping_nop.yaml
@@ -1,4 +1,3 @@
----
 - name: unwrapping
   pull:
     plugin: pnp.plugins.pull.simple.Repeat
