+++ config/readme/08_udf.yaml
@@ -1,4 +1,3 @@
----
 udfs:
   - name: my_str  # Register the builtin str callable as my_str
     plugin: str
