+++ docs/engines/async.yaml
@@ -1,11 +1,11 @@
-# Will use threads to accomplish concurrency
-# Drawback: If a plugin does not stop gracefully the termination will hang...
+# Will use asyncio to accomplish concurrency
+
 engine: !engine
-  type: pnp.engines.async_.AsyncEngine
+  type: pnp.engines.AsyncEngine
   retry_handler: !retry
     type: pnp.engines.SimpleRetryHandler
 tasks:
-  - name: hello_world
+  - name: async
     pull:
       plugin: pnp.plugins.pull.simple.Repeat
       args:
@@ -13,13 +13,3 @@ tasks:
         repeat: "Hello World"
     push:
       - plugin: pnp.plugins.push.simple.Echo
-  - name: counter
-    pull:
-      plugin: pnp.plugins.pull.simple.Count
-      args:
-        wait: 1
-    push:
-      - plugin: pnp.plugins.push.simple.Echo
-        deps:
-          - plugin: pnp.plugins.push.simple.Echo
-            selector: "'{} + 1 = {}'.format(payload, int(payload) + 1)"
