+++ docs/engines/index.mdpp
@@ -1,19 +1,25 @@
 If you do not specify any engine the `ThreadEngine` is chosen by default accompanied by the `AdvancedRetryHandler`.
 This keeps maximum backwards compatibility.
 
-# pnp.engines.sequential.SequentialEngine
+Remark: This will probably change in the near future in favor to the `AsyncEngine`.
+
+# pnp.engines.SequentialEngine
 
 By using the `Sequential` engine you can run your configs as scripts. Given the example below, the "script" will
 end when it has finished counting to 3. Make sure to use the `NoRetryHandler` to actually end the runner when
 the pull has finished instead of retrying the "failed" `pull`. You cn only run a single task not multiple.
-When you want to run multiple task in a concurrent manner you have to use the `ThreadEngine` or the `ProcessEngine`.
+When you want to run multiple task in a concurrent manner you have to use the `ThreadEngine`, `ProcessEngine` or `AsyncEngine`.
 
 !INCLUDECODE "docs/engines/sequential.yaml" (yaml)
 
-# pnp.engines.thread.ThreadEngine
+# pnp.engines.ThreadEngine
 
 !INCLUDECODE "docs/engines/threading.yaml" (yaml)
 
-# pnp.engines.process.ProcessEngine
+# pnp.engines.ProcessEngine
 
 !INCLUDECODE "docs/engines/process.yaml" (yaml)
+
+# pnp.engines.AsyncEngine (0.18.0+)
+
+!INCLUDECODE "docs/engines/async.yaml" (yaml)
