+++ mypy.ini
@@ -1,0 +1,16 @@
+[mypy]
+disallow_any_generics = True
+disallow_subclassing_any = True
+disallow_untyped_calls = True
+disallow_untyped_defs = True
+disallow_incomplete_defs = True
+check_untyped_defs = True
+disallow_untyped_decorators = True
+no_implicit_optional = True
+warn_unused_ignores = False
+warn_no_return = True
+warn_return_any = True
+show_none_errors = True
+allow_redefinition = False
+warn_redundant_casts = True
+warn_unused_configs = True
