+++ pnp/engines/async_.py
@@ -1,12 +1,14 @@
 """Base implementation for asynchronous engines."""
+import time
 from typing import Optional, Any
 
 import asyncio
 
-from . import Engine, RetryHandler, SimpleRetryHandler, PushExecutor
+from .base import Engine, RetryHandler, SimpleRetryHandler, PushExecutor
 from ..models import TaskSet, TaskModel, PushModel
 from ..plugins.pull import AsyncPullBase
 from ..plugins.push import PushBase
+from ..shared.async_ import async_sleep_until_interrupt
 from ..typing import Payload
 
 
@@ -30,13 +32,7 @@ class AsyncEngine(Engine):
         try:
             self.loop.run_until_complete(asyncio.gather(*coros, loop=self.loop))
         except KeyboardInterrupt:
-            self._shutdown()
-        finally:
-            # This check is only needed for Python 3.5 and below
-            if hasattr(self.loop, "shutdown_asyncgens"):
-                self.loop.run_until_complete(self.loop.shutdown_asyncgens())
-            if self.loop.is_running():
-                self.loop.call_soon_threadsafe(self.loop.stop)
+            self.stop()
 
     def _stop(self) -> None:
         if not self.tasks:
@@ -53,7 +49,7 @@ class AsyncEngine(Engine):
                     payload,
                     push
                 )
-                self.loop.create_task(self._schedule_push(payload, push))
+                asyncio.run_coroutine_threadsafe(self._schedule_push(payload, push), self.loop)
 
         task.pull.instance.on_payload = on_payload  # type: ignore
 
@@ -99,10 +95,14 @@ class AsyncEngine(Engine):
                     task.name,
                     task.pull.instance,
                 )
+                task.pull.instance.stop()
             else:
-                await asyncio.sleep(directive.wait_for, loop=self.loop)
+                async def _interrupt() -> bool:
+                    return task.pull.instance.stopped
+                await async_sleep_until_interrupt(directive.wait_for, _interrupt)
 
     async def _stop_task(self, task: TaskModel) -> None:
+        self.logger.info("Stopping task %s", task.name)
         await self.loop.run_in_executor(None, task.pull.instance.stop)
 
     async def _schedule_push(self, payload: Payload, push: PushModel) -> None:
@@ -110,14 +110,20 @@ class AsyncEngine(Engine):
         assert isinstance(push.instance, PushBase)
 
         def _callback(result: Payload, dependency: PushModel) -> None:
-            self.loop.create_task(self._schedule_push(result, dependency))
+            asyncio.run_coroutine_threadsafe(self._schedule_push(result, dependency), self.loop)
 
-        await PushExecutor().async_execute(
-            push.instance.name,
-            payload,
-            push,
-            _callback
-        )
+        try:
+            await PushExecutor().async_execute(
+                push.instance.name,
+                payload,
+                push,
+                _callback
+            )
+        except KeyboardInterrupt:  # pragma: no cover
+            pass
+        except Exception:  # pragma: no cover, pylint: disable=broad-except
+            import traceback
+            self.logger.error("\n%s", traceback.format_exc())
 
     def _shutdown(self) -> None:
         assert self.tasks and isinstance(self.tasks, dict)
@@ -125,7 +131,14 @@ class AsyncEngine(Engine):
         loop = self.loop
 
         # Optionally show a message if the shutdown may take a while
-        self.logger.info("Attempting graceful shutdown, press Ctrl+C again to exit...")
+        self.logger.info("Attempting graceful shutdown, press Ctrl+C again to exit forcefully...")
+
+        # If it's running: Stop it!
+        # Two scenarios:
+        # 1. KeyboardInterrupt/OOM/Other Exception -> loop is already down
+        # 2. stop() called -> We need to stop the loop (the code below assumes that the
+        #   loop is down)
+        self._wait_for_loop_to_stop(loop)
 
         # Do not show `asyncio.CancelledError` exceptions during shutdown
         # a lot of these may be generated, skip this if you prefer to see them
@@ -140,19 +153,30 @@ class AsyncEngine(Engine):
             *[self._stop_task(task) for _, task in self.tasks.items()],
             loop=loop
         )
-        if loop.is_running():
-            loop.create_task(stop_tasks)
-        else:
-            loop.run_until_complete(stop_tasks)
+        loop.run_until_complete(stop_tasks)
 
         # Handle shutdown gracefully by waiting for all tasks to be cancelled
         tasks = asyncio.gather(*asyncio.Task.all_tasks(loop=loop), loop=loop,
                                return_exceptions=True)
         tasks.add_done_callback(lambda t: loop.stop())
-        # tasks.cancel()
 
         # Keep the event loop running until it is either destroyed or all
         # tasks have really terminated
         while not tasks.done() and not loop.is_closed():
             if not loop.is_running():
                 loop.run_forever()
+
+        # This check is only needed for Python 3.5 and below
+        if hasattr(self.loop, "shutdown_asyncgens"):
+            self.loop.run_until_complete(self.loop.shutdown_asyncgens())  # type: ignore
+
+    @staticmethod
+    def _wait_for_loop_to_stop(loop: asyncio.AbstractEventLoop) -> None:
+        if loop.is_running():
+            loop.stop()
+        for _ in range(10):
+            if not loop.is_running():
+                break
+            time.sleep(1)
+        if loop.is_running():
+            raise RuntimeError("Eventloop did not stop")
