+++ tests/manual/async_complex.yaml
@@ -1,0 +1,36 @@
+engine: !engine
+  type: pnp.engines.async_.AsyncEngine
+  retry_handler: !retry
+    type: pnp.engines.NoRetryHandler
+tasks:
+  - name: counter
+    pull:
+      plugin: pnp.plugins.pull.simple.Count
+      args:
+        wait: 1
+    push:
+      - plugin: pnp.plugins.push.simple.Echo
+        deps:
+          - plugin: pnp.plugins.push.simple.Echo
+            selector: "int(payload) * int(payload)"
+  - name: repeat
+    pull:
+      plugin: pnp.plugins.pull.simple.Repeat
+      args:
+        repeat: "hello world"
+    push:
+      - plugin: pnp.plugins.push.simple.Echo
+        selector:
+          head: "lambda p: p.split(' ')[0]"
+          tail: "lambda p: p.split(' ')[1:]"
+  - name: dht
+    pull:
+      plugin: pnp.plugins.pull.sensor.DHT
+      args:
+        interval: 10s
+        instant_run: true
+    push:
+      - plugin: pnp.plugins.push.simple.Echo
+        selector:
+          humidity: "lambda p: p.humidity"
+          temp: "lambda p: p.temperature"
