+++ tests/manual/async_push_error.yaml
@@ -1,14 +1,13 @@
 engine: !engine
-  type: pnp.engines.SequentialEngine
+  type: pnp.engines.async_.AsyncEngine
   retry_handler: !retry
     type: pnp.engines.NoRetryHandler
 tasks:
-  - name: sequential
+  - name: push_error
     pull:
       plugin: pnp.plugins.pull.simple.Count
       args:
-        wait: 1
-        from_cnt: 1
-        to_cnt: 4
+        wait: 2
+        to_cnt: 5
     push:
-      - plugin: pnp.plugins.push.simple.Echo
+      - plugin: tests.manual.plugins.ErroneousPush
