+++ tests/manual/thread_long_poll.yaml
@@ -1,0 +1,21 @@
+engine: !engine
+  type: pnp.engines.ThreadEngine
+  retry_handler: !retry
+    type: pnp.engines.NoRetryHandler
+tasks:
+  - name: blocking_long_poll
+    pull:
+      plugin: tests.manual.plugins.LongPoll
+      args:
+        interval: 5
+        instant_run: true
+    push:
+      - plugin: pnp.plugins.push.simple.Echo
+  - name: async_long_poll
+    pull:
+      plugin: tests.manual.plugins.AsyncLongPoll
+      args:
+        interval: 3
+        instant_run: true
+    push:
+      - plugin: pnp.plugins.push.simple.Echo
